requirements.txt
pytest>=8.4.2
pytest-playwright>=0.7.0
allure-pytest>=2.14.3
playwright